+++ demisto_sdk/commands/common/hook_validations/base_validator.py
@@ -63,11 +63,14 @@ class BaseValidator:
         Returns:
             str. Will return the formatted error message if it is not ignored, an None if it is ignored
         """
-        error_type = 'WARNING' if warning else 'ERROR'
-        formatted_error = f"[{error_type}]: {file_path}: [{error_code}] - {error_message}".rstrip("\n") + "\n"
+        def formatted_error_func(error_type):
+            if error_type not in {'ERROR', 'WARNING'}:
+                raise ValueError("Error type is not valid.")
+            formatted = f"[{error_type}]: {file_path}: [{error_code}] - {error_message}".rstrip("\n") + "\n"
+            if drop_line:
+                formatted = '\n' + formatted
+            return formatted
 
-        if drop_line:
-            formatted_error = "\n" + formatted_error
         if file_path:
             if not isinstance(file_path, str):
                 file_path = str(file_path)
@@ -85,11 +88,12 @@ class BaseValidator:
 
         if self.should_ignore_error(error_code, ignored_errors) or warning:
             if self.print_as_warnings or warning:
-                click.secho(formatted_error, fg="yellow")
+                click.secho(formatted_error_func('WARNING'), fg="yellow")
                 self.json_output(file_path, error_code, error_message, warning)
                 self.add_to_report_error_list(error_code, file_path, FOUND_FILES_AND_IGNORED_ERRORS)
             return None
 
+        formatted_error = formatted_error_func('ERROR')
         if should_print and not self.suppress_print:
             if suggested_fix:
                 click.secho(formatted_error[:-1], fg="bright_red")